% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\docType{data}
\name{BC_lakes}
\alias{BC_lakes}
\title{Dataset from 256 surveys of small lakes in British Columbia, Canada}
\description{
Dataset from 256 surveys of small lakes in British Columbia, Canada
}
\examples{
data(BC_lakes)

plot(BC_lakes[[1]]) # Plot data

mod <- fit_model(BC_lakes[[1]]) # Run model
plot(mod) # View model results
}
