% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{fit_model}
\alias{fit_model}
\title{Fit the model}
\usage{
fit_model(RBdata, start = NULL, nit_harvest = 5L, use_priors = TRUE,
  lower = NULL, upper = NULL, control = list(iter.max = 1000,
  eval.max = 1000), ...)
}
\arguments{
\item{RBdata}{An object of class \linkS4class{RBdata} containing data and priors for the model.}

\item{start}{An optional list of starting values. Otherwise, the prior means are used as starting values.}

\item{nit_harvest}{The number of iterations for solving F and the probability of harvesting in the model.}

\item{use_priors}{Logical, whether to use priors for the parameters in the model.}

\item{lower}{A vector of the lower bounds of the parameters. This argument overrides the default, and thus generally not recommended to use.}

\item{upper}{A vector of the upper bounds of the parameters. This argument overrides the default, and thus generally not recommended to use.}

\item{control}{A list of control arguments to be passed to \link[stats]{nlminb}.}

\item{...}{More arguments to pass in.}
}
\description{
Setup the TMB model and runs the optimization to return the posterior mode.
}
\examples{
data(BC_lakes)
mod <- fit_model(BC_lakes[[1]]) # Run model
plot(mod) # View model results
}
\seealso{
\link{run_mcmc} \link{simulation}
}
