% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_definitions.R
\docType{methods}
\name{plot,RBdata,ANY-method}
\alias{plot,RBdata,ANY-method}
\title{\code{plot} method for S4 class \code{RBdata}}
\usage{
\S4method{plot}{RBdata,ANY}(x, bubble = 7, data = TRUE, prior = TRUE,
  ...)
}
\arguments{
\item{x}{An object of class \linkS4class{RBdata}.}

\item{bubble}{A number for adjusting the bubble size in the plot for age-length data.}

\item{data}{Logical, whether a figure showing the data will be plotted.}

\item{prior}{Logical, whether a figure showing priors will be plotted.}

\item{...}{Other miscellaneous variables (not currently used).}
}
\description{
Plots length frequency and age-length data, as well as priors.
}
