% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_definitions.R
\docType{methods}
\name{plot,RBfit,missing-method}
\alias{plot,RBfit,missing-method}
\title{\code{plot} method for S4 class \code{RBfit}}
\usage{
\S4method{plot}{RBfit,missing}(x, bubble = 7, diagnostic = TRUE,
  posterior = TRUE, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{RBfit}.}

\item{bubble}{A number for adjusting the bubble size in the plot for age-length data.}

\item{diagnostic}{Logical, whether a figure of model diagnostics (residuals, model fits, etc.) will be plotted.}

\item{posterior}{Logical, whether a figure of posterior distributions will be plotted.}

\item{...}{Other miscellaneous variables (not currently used).}
}
\description{
Plots length frequency and age-length data, as well as priors.
}
