% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_definitions.R
\docType{methods}
\name{plot,stanfit,missing-method}
\alias{plot,stanfit,missing-method}
\alias{plot,stanfit,stanfit-method}
\title{\code{plot} method for S4 class \code{stanfit}}
\usage{
\S4method{plot}{stanfit,missing}(x, y, ...)

\S4method{plot}{stanfit,stanfit}(x, y, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{stanfit}.}

\item{y}{Another object of class \linkS4class{stanfit} which only sampled the priors.}
}
\description{
Returns a figure parameter priors and posteriors. See example from \link{run_mcmc}.
}
\examples{
\donttest{
data(BC_lakes)
mod <- fit_model(BC_lakes[[1]]) # Run model

pr_only <- run_mcmc(mod, priors_only = TRUE) # Run MCMC on priors
plot(pr_only) # Shows density plots of priors

samps <- run_mcmc(mod) # Run MCMC with likelihood for data
plot(samps) # Shows density plots of posteriors
plot(samps, pr_only) # Shows density plots of both priors and posteriors

##### Any plotting diagnostic function from the rstan package can be used.
summary(samps) # Summary of parameter values
rstan::stan_ac(samps) # Autocorrelation plot of MCMC chains
rstan::stan_trace(samps) # Trace plot of MCMC chains
rstan::stan_dens(samps) # Density plot of posteriors
}
}
\seealso{
\link{run_mcmc}
}
