% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\docType{methods}
\name{plot.RBdata}
\alias{plot.RBdata}
\alias{plot,RBfit,missing-method}
\alias{plot.RBfit}
\alias{plot,stanfit,missing-method}
\alias{plot.stanfit}
\alias{plot,stanfit,stanfit-method}
\title{\code{plot} methods for S4 classes in \code{RBassess} package}
\usage{
\S4method{plot}{RBdata,ANY}(x, bubble = 7, data = TRUE, prior = TRUE,
  ...)

\S4method{plot}{RBfit,missing}(x, bubble = 7, diagnostic = TRUE,
  posterior = TRUE, ...)

\S4method{plot}{stanfit,missing}(x, y, ...)

\S4method{plot}{stanfit,stanfit}(x, y, ...)
}
\arguments{
\item{x}{An object of class \linkS4class{RBdata}, \linkS4class{RBfit}, or \linkS4class{stanfit}.}

\item{bubble}{A number for adjusting the bubble size in the plot for age-length data.}

\item{data}{Logical, whether a figure showing the data will be plotted.}

\item{prior}{Logical, whether a figure showing priors will be plotted.}

\item{...}{Other miscellaneous variables (not currently used).}

\item{diagnostic}{Logical, whether a figure of model diagnostics (residuals, model fits, etc.) will be plotted.}

\item{posterior}{Logical, whether a figure of posterior distributions will be plotted.}

\item{y}{Optional, another object of class \linkS4class{stanfit} which only sampled the priors.}
}
\description{
A suite of plot functions for various objects in the RBassess packages. For data, plots length frequency and age-length data, as well as priors.
For posterior mode and MCMC model results, plots model fits and posteriors.
}
\examples{
###### View data
data(BC_lakes)
plot(BC_lakes[[1]])

###### Run model
mod <- fit_model(BC_lakes[[1]])
plot(mode)

\donttest{
###### Run MCMC
pr_only <- run_mcmc(mod, priors_only = TRUE) # Run MCMC on priors
plot(pr_only) # Shows density plots of priors

samps <- run_mcmc(mod) # Run MCMC with likelihood for data

plot(samps) # Shows density plots of posteriors
plot(samps, pr_only) # Shows density plots of both priors and posteriors

##### Any plotting diagnostic function from the rstan package are also available.
summary(samps) # Summary of parameter values
rstan::stan_ac(samps) # Autocorrelation plot of MCMC chains
rstan::stan_trace(samps) # Trace plot of MCMC chains
rstan::stan_dens(samps) # Density plot of posteriors
}
}
\seealso{
\link{fit_model} \link{run_mcmc}
}
