% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_definitions.R
\docType{class}
\name{RBdata-class}
\alias{RBdata-class}
\alias{RBdata}
\title{Class-\code{RBdata}}
\description{
An S4 class for setting up inputs for the rainbow trout assessment.
}
\section{Slots}{

\describe{
\item{\code{Lake}}{Name of lake.}

\item{\code{Year}}{Year of survey.}

\item{\code{Length_bin}}{A vector of midpoints of the length bins. Width of all length bins assumed to be equal. Typically in centimeters.}

\item{\code{Age}}{A vector of integer ages. Should be consecutive from 1 to the maximum age.}

\item{\code{Age_adjust}}{A vector of accumulated growing degree days (converted to years) corresponding to the integer ages.}

\item{\code{Length}}{A vector (single year sample) for the length composition data. The vector should be of
length equal to \code{length(Length_bins)}.}

\item{\code{Age_length}}{A matrix (single year sample) for the age-length sammples. The matrix should have
\code{length(Age)} rows and \code{length(Length_bins)} columns.}

\item{\code{L_stock}}{A vector of mean lengths at stocking, in order of the corresponding ages.}

\item{\code{stock_density}}{A vector of stocking density, in order of the corresponding ages. In units of numbers per hectacre.}

\item{\code{bag_limit}}{The daily bag limit (number of fish per angler day) used to calculated the probability of harvesting.}

\item{\code{M_release}}{The assumed instantaneous mortality rate (per year) from catch and release.}

\item{\code{prior_Linf}}{Von Bertalanffy asymptotic length. Vector of length two for mean and standard deviation, respectively.}

\item{\code{prior_K}}{Von Bertalanffy growth coefficient. Vector of length two for mean and standard deviation, respectively.}

\item{\code{prior_CV_Len}}{Coefficient in variation in length-at-age. Vector of length two for mean and standard deviation, respectively.}

\item{\code{prior_M}}{Instantaneous natural mortality (per year). Vector of length two for mean and standard deviation, respectively.}

\item{\code{prior_Effort}}{Fishing effort. Vector of length two for mean and standard deviation, respectively. In units of angler days per hectacre.}

\item{\code{prior_q}}{Catchability coefficient for scaling effort. Vector of length two for mean and standard deviation, respectively.}

\item{\code{prior_GN_SL50}}{Gillnet length at 50\% selectivity for the survey. Vector of length two for mean and standard deviation, respectively.
Default priors are based on values from Askey et al. (2007).}

\item{\code{prior_GN_gamma}}{Steepness of selectivity for gillnet (positive values). Vector of length two for mean and standard deviation, respectively.
Default priors are based on values from Askey et al. (2007).}

\item{\code{prior_angler_gamma}}{Angler length at 50\% selectivity. Vector of length two for mean and standard deviation, respectively.}

\item{\code{prior_angler_SL50}}{Steepness of selectivity for angler (positive values). Vector of length two for mean and standard deviation, respectively.}
}}

\examples{
new_data <- new("RBdata")
}
\references{
Asky, P.J. Post, J.R., Parkinson, E.A., Rivot, E., Paul, A.J., and Biro, P.A. 2007. Estimation of gillnet efficiency and selectivity across multiple
sampling units: A hierarchical Bayesian analysis using mark-recapture data. Fisheries Research 83: 162-174.
}
\seealso{
\link{fit_model} \link{run_mcmc} \link{BC_lakes} \link{plot.RBdata}
}
