% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_definitions.R
\docType{class}
\name{RBfit-class}
\alias{RBfit-class}
\alias{RBfit}
\title{Class-\code{RBfit}}
\description{
An S4 class for posterior mode output from rainbow trout assessment model.
}
\section{Slots}{

\describe{
\item{\code{obj}}{The TMB assessment object.}

\item{\code{opt}}{Output from \link[stats]{nlminb}.}

\item{\code{SD}}{Output from \link[TMB]{sdreport}.}

\item{\code{report}}{Output from TMB model.}

\item{\code{RBdata}}{Data objected used to generate the assessment.}
}}

\seealso{
\link{plot,RBfit,missing-method} \link{summary,RBfit-method}
}
