% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Run MCMC}
\usage{
run_mcmc(RBfit, priors_only = FALSE, chains = 2L, iter = 20000,
  warmup = 0.5 * iter, thin = 5, seed = 1, cores = chains,
  lower = NULL, upper = NULL, ...)
}
\arguments{
\item{RBfit}{An object of class \linkS4class{RBfit}.}

\item{priors_only}{If \code{TRUE}, turns off the likelihood and samples the priors.}

\item{chains}{The number of chains for running the MCMC.}

\item{iter}{The number of total iterations of the MCMC.}

\item{warmup}{The number of burn-in iterations of the MCMC.}

\item{thin}{How often the iterations are saved.}

\item{seed}{An integer for indexing random number generation.}

\item{cores}{The number of CPU cores for running MCMC in parallel.}

\item{lower}{A vector of the lower bounds of the parameters. This argument overrides the default, and thus generally not recommended to use.}

\item{upper}{A vector of the upper bounds of the parameters. This argument overrides the default, and thus generally not recommended to use.}

\item{...}{More arguments to pass to \code{rstan::sampling} via \link[tmbstan]{tmbstan}, for example, \code{init} for starting values for the MCMC.}
}
\description{
Set up a stan model and runs the MCMC from a fitted TMB model (see example).
}
\examples{
\donttest{
data(BC_lakes)
mod <- fit_model(BC_lakes[[1]]) # Run model

pr_only <- run_mcmc(mod, priors_only = TRUE) # Run MCMC on priors
plot(pr_only) # Shows density plots of priors

samps <- run_mcmc(mod) # Run MCMC with likelihood for data
plot(samps) # Shows density plots of posteriors
plot(samps, pr_only) # Shows density plots of both priors and posteriors

##### Any plotting diagnostic function from the rstan package can be used.
summary(samps) # Summary of parameter values
stan_ac(samps) # Autocorrelation plot of MCMC chains
stan_trace(samps) # Trace plot of posteriors
stan_dens(samps) # Density plot of posteriors
}
}
\seealso{
\link{fit_model} \link{plot,stanfit,missing-method}
}
