% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_definitions.R
\docType{methods}
\name{summary,RBfit-method}
\alias{summary,RBfit-method}
\title{\code{summary} method for S4 class \code{RBfit}}
\usage{
\S4method{summary}{RBfit}(object, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{RBfit}.}
}
\description{
Returns parameter estimates and standard deviations.
}
