% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4_methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary.RBfit}
\alias{summary,stanfit-method}
\alias{summary.stanfit}
\title{\code{summary} method for S4 class \code{RBfit}}
\usage{
\S4method{summary}{RBfit}(object, description = FALSE, digits = 2, ...)

\S4method{summary}{stanfit}(object, description = FALSE, digits = 2,
  full = FALSE, ...)
}
\arguments{
\item{object}{An object of class \linkS4class{RBfit} or \linkS4class{stanfit}.}

\item{description}{Logical, indicating whether an additional column with parameter descriptions.}

\item{digits}{The number of decimal places for rounding. Use \code{NA} for no rounding.}

\item{...}{Additional arguments}

\item{full}{Logical, returns more information for a stanfit object: posterior quantiles, Rhat, and effective sample size.}
}
\value{
A data frame.
}
\description{
Returns parameter estimates and standard deviations.
}
\examples{
data(BC_lakes)
res <- fit_model(BC_lakes[[69]])
summary(res)

\donttest{
samps <- run_mcmc(res)
summary(samps)
}
}
