% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\docType{methods}
\name{report}
\alias{report}
\title{HTML report for assessment model}
\usage{
\S4method{report}{RBfit,missing}(x, dir = tempdir(), author = NULL,
  bubble = 7, ...)
}
\arguments{
\item{x}{An object of class `RBfit`, returned from \link{fit_model}.}

\item{dir}{The directory for saving the HTML document.}

\item{author}{A character string for the author.}
}
\description{
Returns an HTML document with figures showing data, model fits, and diagnostics.
}
