% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{simulation}
\alias{simulation}
\title{Generates stochastic length and age-length samples runs a simulation}
\usage{
simulation(RBfit, pars, N = c(50, 100), seed = 1)
}
\arguments{
\item{RBfit}{A fitted assessment object used for setting up the simulation.}
}
\description{
Simulate data for assessment model.
}
